+++ src/evm/EvmStateMachine.ts
@@ -7,7 +7,7 @@ import {
 import { TransactionStruct } from "@typechain-types/contracts/V1/DataTypes";
 import StateManager from "@/stateManager/StateManager";
 import Clock from "@/Clock";
-import { TimeConfig } from "@/DataTypes";
+import { TimeConfig } from "@/types";
 import DebugProxy from "@/utils/DebugProxy";
 import P2pEventHooks from "@/P2pEventHooks";
 import AStateMachine from "@/AStateMachine";
