+++ src/types/time.ts
@@ -1,0 +1,6 @@
+export type TimeConfig = {
+    p2pTime: number;
+    agreementTime: number;
+    chainFallbackTime: number;
+    challengeTime: number;
+};
